[MIG] shopinvader_api_address
