+++ shopinvader_api_address/README.rst
@@ -1,9 +1,5 @@
-.. image:: https://odoo-community.org/readme-banner-image
-   :target: https://odoo-community.org/get-involved?utm_source=readme
-   :alt: Odoo Community Association
-
 =======================
-Shopinvader Address Api
+Shopinvader API Address
 =======================
 
 .. 
@@ -17,17 +13,17 @@ Shopinvader Address Api
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
     :target: https://odoo-community.org/page/development-status
     :alt: Beta
-.. |badge2| image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-shopinvader%2Fodoo--shopinvader-lightgray.png?logo=github
-    :target: https://github.com/shopinvader/odoo-shopinvader/tree/16.0/shopinvader_api_address
+    :target: https://github.com/shopinvader/odoo-shopinvader/tree/18.0/shopinvader_api_address
     :alt: shopinvader/odoo-shopinvader
 
 |badge1| |badge2| |badge3|
 
-This addon adds a web API on top of the res.partner model to ease the management
-of addresses from Web frontend.
+This addon adds a web API on top of the res.partner model to ease the
+management of addresses from Web frontend.
 
 **Table of contents**
 
@@ -40,7 +36,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/shopinvader/odoo-shopinvader/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/shopinvader/odoo-shopinvader/issues/new?body=module:%20shopinvader_api_address%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/shopinvader/odoo-shopinvader/issues/new?body=module:%20shopinvader_api_address%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -48,21 +44,21 @@ Credits
 =======
 
 Authors
-~~~~~~~
+-------
 
 * ACSONE SA/NV
 
 Contributors
-~~~~~~~~~~~~
+------------
 
-* Laurent Mignon <laurent.mignon@acsone.eu>
-* Marie Lejeune <marie.lejeune@acsone.eu>
-* Stéphane Bidoul <stephane.bidoul@acsone.eu>
-* Zina Rasoamanana <zina.rasoamanana@acsone.eu>
+- Laurent Mignon <laurent.mignon@acsone.eu>
+- Marie Lejeune <marie.lejeune@acsone.eu>
+- Stéphane Bidoul <stephane.bidoul@acsone.eu>
+- Zina Rasoamanana <zina.rasoamanana@acsone.eu>
 
 Maintainers
-~~~~~~~~~~~
+-----------
 
-This module is part of the `shopinvader/odoo-shopinvader <https://github.com/shopinvader/odoo-shopinvader/tree/16.0/shopinvader_api_address>`_ project on GitHub.
+This module is part of the `shopinvader/odoo-shopinvader <https://github.com/shopinvader/odoo-shopinvader/tree/18.0/shopinvader_api_address>`_ project on GitHub.
 
 You are welcome to contribute.
