+++ shopinvader_api_address/__manifest__.py
@@ -2,26 +2,20 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 {
-    "name": "Shopinvader Address Api",
-    "summary": """
-        Adds a service to manage shopinvader invoicing and delivery address""",
-    "version": "16.0.1.3.3",
+    "name": "Shopinvader API Address",
+    "summary": "Adds a service to manage shopinvader invoicing and delivery address",
+    "version": "18.0.1.0.0",
     "license": "AGPL-3",
     "author": "ACSONE SA/NV,Odoo Community Association (OCA)",
     "website": "https://github.com/shopinvader/odoo-shopinvader",
     "depends": [
-        "extendable",
-        "extendable_fastapi",
-        "fastapi",
-        "shopinvader_address",
+        "shopinvader_router_helper",
         "shopinvader_schema_address",
+        "extendable_fastapi",
     ],
     "data": [
         "security/res_groups.xml",
         "security/res_partner.xml",
     ],
-    "external_dependencies": {
-        "python": ["fastapi", "extendable_pydantic>=1.0.0", "pydantic>=2.0.0"]
-    },
-    'installable': False,
+    "installable": True,
 }
