+++ shopinvader_api_address/readme/CONTRIBUTORS.md
@@ -1,0 +1,4 @@
+- Laurent Mignon \<laurent.mignon@acsone.eu\>
+- Marie Lejeune \<marie.lejeune@acsone.eu\>
+- Stéphane Bidoul \<stephane.bidoul@acsone.eu\>
+- Zina Rasoamanana \<zina.rasoamanana@acsone.eu\>
