--- shopinvader_api_address/readme/CONTRIBUTORS.rst
@@ -1,4 +1,0 @@
-* Laurent Mignon <laurent.mignon@acsone.eu>
-* Marie Lejeune <marie.lejeune@acsone.eu>
-* Stéphane Bidoul <stephane.bidoul@acsone.eu>
-* Zina Rasoamanana <zina.rasoamanana@acsone.eu>
