--- shopinvader_api_address/readme/DESCRIPTION.rst
@@ -1,2 +1,0 @@
-This addon adds a web API on top of the res.partner model to ease the management
-of addresses from Web frontend.
