+++ shopinvader_api_address/routers/__init__.py
@@ -1,2 +1,2 @@
-from .address_service import address_router
-from . import address_service
+from .addresses import address_router
+from . import addresses
