+++ shopinvader_api_address/security/res_groups.xml
@@ -2,15 +2,13 @@
 <!-- Copyright 2023 ACSONE SA/NV
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
 <odoo>
-  <record id="shopinvader_address_user_group" model="res.groups">
-    <field name="name">Shopinvader Address user</field>
-    <field
+    <record id="shopinvader_address_user_group" model="res.groups">
+        <field name="name">Shopinvader Address user</field>
+        <field
             name="implied_ids"
             eval="[
       (4, ref('fastapi.group_fastapi_endpoint_runner')),
       ]"
         />
-
-  </record>
-
+    </record>
 </odoo>
