+++ shopinvader_api_address/security/res_partner.xml
@@ -11,7 +11,6 @@
             eval="[(4, ref('shopinvader_api_address.shopinvader_address_user_group'))]"
         />
         <field name="domain_force">[('parent_id','=',authenticated_partner_id)]</field>
-
     </record>
 
     <record id="shopinvader_address_res_partner_access" model="ir.model.access">
@@ -23,5 +22,4 @@
         <field name="perm_create" eval="True" />
         <field name="perm_unlink" eval="False" />
     </record>
-
 </odoo>
